requirements.txt
# نظام تقسيط - Python + PostgreSQL
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator>=2.2.0
WTForms>=3.0
python-dateutil>=2.8
